(#30) : enable (:) char at keys
__WHY__
See #30

IRIs (Internationalized Resource Identifiers [RFC3987]) are fundamental
to Linked Data as that is how most nodes and properties are identified.
A compact IRI is a way of expressing an IRI using a prefix and suffix
separated by a colon (:).

For example, foaf:name would be expanded to the IRI http://xmlns.com/foaf/0.1/name.
```
{
  "@context":
  {
    "foaf": "http://xmlns.com/foaf/0.1/"
...
  },
  "@type": "foaf:Person"
  "foaf:name": "Dave Longley",
...
}
```

__WHAT__
* fix regex pattern to support (:) at name
* update tests to validate a new character
* bump lib version and update distribution